refactor: Change service logic structure
